[rocprofiler-compute] Move the marker-CSV tests to test_utils_profile.py
An audit of every unit file against the module its name claims turned up
one more case of the same placement error: test_rocpd_data.py held tests
for _parse_function_backend and _augment_marker_csv, both in
utils/utils_profile.py.

The process_ml_api_trace_output tests stay put despite naming
utils_analysis; they share the COUNTER_ROWS fixture with the rocpd_data
tests, so moving them would duplicate it.

Bodies are unchanged, so node IDs are identical before and after.
